Refactor results factory
