+++ knotx-core/src/main/java/com/cognifide/knotx/repository/RemoteRepository.java
@@ -17,7 +17,7 @@
  */
 package com.cognifide.knotx.repository;
 
-import com.cognifide.knotx.util.AsyncResultFactory;
+import com.cognifide.knotx.result.AsyncResultFactory;
 
 import org.springframework.http.HttpStatus;
