+++ knotx-core/src/main/java/com/cognifide/knotx/result/FailureAsyncResult.java
@@ -1,0 +1,49 @@
+/*
+ * Knot.x - Reactive microservice assembler
+ *
+ * Copyright (C) 2016 Cognifide Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.cognifide.knotx.result;
+
+import io.vertx.core.AsyncResult;
+
+public class FailureAsyncResult<T> implements AsyncResult<T> {
+
+    private final Throwable cause;
+
+    public FailureAsyncResult(Throwable cause) {
+        this.cause = cause;
+    }
+
+    @Override
+    public T result() {
+        return null;
+    }
+
+    @Override
+    public Throwable cause() {
+        return cause;
+    }
+
+    @Override
+    public boolean succeeded() {
+        return false;
+    }
+
+    @Override
+    public boolean failed() {
+        return true;
+    }
+}
